Finish GS switch to mpreal; still needs testing

+++ test.cpp
@@ -186,7 +186,7 @@ bool UPlusIntegral(OStream& console) {
 bool UPlusIntegral_Case(const builtin_class a, const builtin_class b, 
         const builtin_class expected, OStream& console) {
     constexpr builtin_class tol = 1e-5;
-    builtin_class answer = ::MatrixInternal::UPlusIntegral(a, b);
+    builtin_class answer = ::MatrixInternal::UPlusIntegral(a, b).toDouble();
     console << "UPlusIntegral(" << a << ", " << b << ") == " << answer;
     if (std::abs(answer - expected) <= tol*answer) {
         console << " == " << expected << " (PASS)" << endl;
